backend/requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
django-cors-headers==3.7.0
drf-yasg==1.20.0
<<<<<<< HEAD
PyJWT == 1.7.1
djangorestframework-simplejwt == 4.4.0
python-environ == 0.4.54
pillow===8.1.2
=======
pillow===8.1.2
django-environ==0.4.5
PyJWT==1.7.1
djangorestframework-simplejwt==4.4.0
>>>>>>> origin/feat/feed
